added logging
commented out what I added so you guys can check if it's correct. I can't run the game because of compilation errors, so I have not tested it already and checked the csv file. I can do this and clean up the csv file once I fix the compilation errors.